requirements.txt
asgiref==3.2.10
pytz==2020.1
sqlparse==0.2.4
coverage==5.2.1
psycopg2-binary==2.8.6
dj-database-url==0.5.0
whitenoise==5.2.0
gunicorn==20.0.4
firebase-admin==5.0.0
pytest==7.1.2
pytest-cov==2.11.1
optimizely-sdk==3.8.0
pymongo==4.2.0
dnspython==2.1.0
requests==2.25.1
drf-yasg==1.20.0
python-dotenv==0.15.0
rest-meets-djongo==0.0.10
marshmallow==3.11.1
dependency-injector==4.31.2
uvicorn[standard]==0.18.2
fastapi==0.79.0
pydantic==1.9.1
email-validator==1.2.1
PyPubSub==4.0.3
